+++ src/allpath_agent/models/__init__.py
@@ -1,3 +1,18 @@
+from .messages import ChatMessage, ChatRequest, ChatResponse, ToolCall
 from .router import ModelProfile, ModelRouter, RoutingDecision, TaskSignals
+from .provider import ChatProvider, FakeProvider, OpenAICompatibleProvider, ProviderError
 
-__all__ = ["ModelProfile", "ModelRouter", "RoutingDecision", "TaskSignals"]
+__all__ = [
+    "ChatProvider",
+    "ChatMessage",
+    "ChatRequest",
+    "ChatResponse",
+    "FakeProvider",
+    "ModelProfile",
+    "ModelRouter",
+    "OpenAICompatibleProvider",
+    "ProviderError",
+    "RoutingDecision",
+    "TaskSignals",
+    "ToolCall",
+]